refactor: components/layout directory for layout related helper widgets
